+++ crates/ruma-events/src/lib.rs
@@ -139,6 +139,8 @@ pub mod macros {
 
 #[cfg(feature = "unstable-msc3927")]
 pub mod audio;
+#[cfg(feature = "unstable-msc3489")]
+pub mod beacon;
 pub mod call;
 pub mod direct;
 pub mod dummy;